Fix Cask format for the latest requirements
Seems that Homebrew started to enforce that audit passes, at least for
the Cask name only including allowed characters.

So rename the Casks, and fix the DSL.

This addresses #217, but `chtf` changes are still needed to support new
Cask names.

For now, leave the old Casks still in the repo to possibly support
older Brew (and chtf) versions.